requirements.txt
absl-py==0.11.0
aiocontextvars==0.2.2
altgraph==0.17
appdirs==1.4.4
asn1crypto==1.3.0
astor==0.8.0
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
bleach==3.1.5
blinker==1.4
bm3d==3.0.7
cached-property==1.5.2
cachetools==3.1.1
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
cloudpickle==1.3.0
colorama==0.4.3
colorit==0.1.0
colormath==3.0.0
contextvars==2.4
cryptography==2.8
cycler==0.10.0
cytoolz==0.10.1
dask==2.13.0
decorator==4.4.2
defusedxml==0.6.0
distlib==0.3.1
easydict==1.9
EasyProcess==0.3
entrypoint2==0.2.1
entrypoints==0.3
ffmpeg==1.4
filelock==3.0.12
Flask==1.1.2
flatbuffers==1.12
fsleyes==0.32.3
fsleyes-props==1.7.0
fsleyes-widgets==0.8.4
fslpy==2.8.4
future==0.18.2
gast==0.2.2
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.34.1
gym==0.17.3
gzip-reader==0.1
h5py==2.10.0
idna==2.9
imageai==2.1.5
imageio==2.8.0
imgaug==0.4.0
immutables==0.14
importlib-metadata==1.6.0
importlib-resources==5.1.0
indexed-gzip==1.2.0
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
isodate==0.6.0
itsdangerous==1.1.0
jedi==0.17.0
Jinja2==2.11.2
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.3
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
keras-facenet==0.3.1
Keras-Preprocessing==1.1.2
keras-vis==0.4.1
kiwisolver==1.1.0
loguru==0.5.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
mss==6.0.0
mtcnn==0.1.0
nbconvert==5.6.1
nbformat==5.0.6
networkx==2.4
nibabel==3.1.0
notebook==6.0.3
numpy==1.19.5
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.34
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.3
pandocfilters==1.4.2
ParallelDots==3.2.14
parso==0.7.0
pathlib==1.0.1
pathlib2==2.3.5
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.1.2
progressbar2==3.51.3
prometheus-client==0.8.0
prompt-toolkit==3.0.5
protobuf==3.11.4
py-git==0.1
pyasn1==0.4.8
pyasn1-modules==0.2.7
pycparser==2.20
pydicom==2.0.0
pyglet==1.5.0
Pygments==2.6.1
PyInstaller==3.6
PyJWT==1.7.1
PyOpenGL==3.1.5
pyOpenSSL==19.1.0
pyparsing==2.4.6
Pypubsub==4.0.3
pyreadline==2.1
pyrsistent==0.16.0
pyscreenshot==2.2
PySocks==1.7.1
python-dateutil==2.8.1
python-utils==2.4.0
pytz==2020.1
PyWavelets==1.1.1
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
PyYAML==5.3
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
Rtree==0.9.4
scikit-image==0.16.2
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
Shapely==1.7.0
six==1.15.0
sklearn==0.0
sphinx-rtd-theme==0.4.3
tb-nightly==2.5.0a20210113
tensorboard==2.1.1
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.1.0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-estimator-nightly==2.5.0.dev2021011301
tf-keras-vis==0.5.3
tf-nightly==2.5.0.dev20210129
tf-slim==1.1.0
tifffile==2020.6.3
tk==0.1.0
toolz==0.10.0
tornado==6.0.4
tqdm==4.43.0
traitlets==4.3.3
trimesh==3.6.43
typing-extensions==3.7.4.3
urllib3==1.25.8
virtualenv==20.4.0
wcwidth==0.1.9
webencodings==0.5.1
Werkzeug==1.0.1
win-inet-pton==1.1.0
win32-setctime==1.0.3
wincertstore==0.2
wrapt==1.12.1
wxnatpy==0.3.2
wxPython==4.0.7.post2
xlrd==1.2.0
XlsxWriter==1.2.8
xnat==0.3.22
zipfile36==0.1.3
zipp==3.1.0
